refactor: ingest cli mostly there
